requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.32.0
httpx~=0.28.0
chromadb~=0.6.0
sentence-transformers~=3.4.0
aiosqlite~=0.20.0
pydantic~=2.10.0
pydantic-settings~=2.7.0
python-dotenv~=1.0.0
sse-starlette~=2.2.0
